@@ -1,69 +1,62 @@
 {
-  "name": "sherpa_frontend",
-  "version": "1.0.0",
-  "description": "Sherpa iOS App React Native",
-  "main": "index.ios.js",
-  "dependencies": {
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-react-native": "^1.9.0",
-    "geo-viewport": "^0.2.1",
-    "install": "^0.8.2",
-    "jquery": "^3.1.0",
-    "lodash": "^4.17.1",
-    "moment": "^2.14.1",
-    "npm": "^4.0.2",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-native": "^0.38.0",
-    "react-native-activity-view": "^0.2.10",
-    "react-native-autogrow-textinput": "^3.0.1",
-    "react-native-camera": "git+https://github.com/lwansbrough/react-native-camera.git",
-    "react-native-camera-roll-picker": "^1.1.8",
-    "react-native-communications": "^2.0.0",
-    "react-native-device-info": "^0.9.3",
-    "react-native-dismiss-keyboard": "^1.0.0",
-    "react-native-enhanced-listview": "^0.10.1",
-    "react-native-exif": "francisco-sanchez-molina/react-native-exif",
-    "react-native-fetch-blob": "^0.9.6",
-    "react-native-gifted-listview": "^0.0.15",
-    "react-native-gifted-spinner": "^0.1.0",
-    "react-native-google-analytics-bridge": "^4.0.1",
-    "react-native-google-places-autocomplete": "^1.2.3",
-    "react-native-hyperlink": "0.0.6",
-    "react-native-i18n": "^0.0.8",
-    "react-native-image-progress": "^0.6.0",
-    "react-native-image-resizer": "0.0.11",
-    "react-native-localization": "^0.1.15",
-    "react-native-mapbox-gl": "^4.1.1",
-    "react-native-maps": "^0.13.0",
-    "react-native-motion-manager": "0.0.6",
-    "react-native-notifications": "^1.0.2",
-    "react-native-orientation": "^1.17.0",
-    "react-native-paged-scroll-view": "^2.0.4",
-    "react-native-pan-controller": "git://github.com/lelandrichardson/react-native-pan-controller.git#emroot/react-0.30",
-    "react-native-progress": "^3.0.1",
-    "react-native-safari-view": "^1.0.1",
-    "react-native-simple-auth": "^0.3.0",
-    "react-native-simple-store": "^0.1.0",
-    "react-native-social-share": "^0.8.0",
-    "react-native-swiper": "^1.4.6",
-    "react-native-tab-navigator": "^0.3.2",
-    "react-native-uploader": "0.0.9",
-    "react-redux": "^3.1.2",
-    "react-router": "^2.5.2",
-    "redux": "^3.5.2",
-    "redux-actions": "^0.8.0",
-    "redux-logger": "^2.6.1",
-    "redux-thunk": "^1.0.3",
-    "supercluster": "^2.2.0",
-    "underscore": "^1.8.3",
-    "whatwg-fetch": "^1.0.0"
-  },
-  "devDependencies": {},
+  "name": "Sherpa",
+  "version": "0.0.1",
+  "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "test": "jest",
+    "postinstall": "scripts/postinstall.sh"
   },
-  "author": "",
-  "license": "ISC"
+  "dependencies": {
+    "exif-js": "^2.2.1",
+    "react": "16.0.0-alpha.6",
+    "react-addons-transition-group": "^15.5.2",
+    "react-native": "0.44.0",
+    "react-native-activity-view": "^0.2.11",
+    "react-native-autogrow-textinput": "^3.0.2",
+    "react-native-blur": "^3.0.0-alpha",
+    "react-native-communications": "^2.2.1",
+    "react-native-deprecated-custom-components": "^0.1.0",
+    "react-native-device-info": "^0.10.2",
+    "react-native-dismiss-keyboard": "^1.0.0",
+    "react-native-exif": "^0.1.0",
+    "react-native-fabric": "^0.4.1",
+    "react-native-fabric-crashlytics": "^0.1.8",
+    "react-native-fetch-blob": "^0.10.5",
+    "react-native-gifted-listview": "0.0.15",
+    "react-native-gifted-spinner": "^0.1.0",
+    "react-native-google-analytics-bridge": "^5.0.1",
+    "react-native-google-places-autocomplete": "^1.2.12",
+    "react-native-hyperlink": "0.0.9",
+    "react-native-image-progress": "^0.6.0",
+    "react-native-image-resizer": "^0.1.0",
+    "react-native-maps": "^0.14.0",
+    "react-native-notifications": "^1.1.10",
+    "react-native-paged-scroll-view": "^2.0.4",
+    "react-native-progress": "^3.2.0",
+    "react-native-safari-view": "^2.0.0",
+    "react-native-sglistview": "^0.4.3",
+    "react-native-simple-auth": "^2.1.0",
+    "react-native-simple-store": "^1.1.0",
+    "react-native-swiper": "^1.5.4",
+    "react-native-tab-navigator": "^0.3.3",
+    "react-redux": "^5.0.4",
+    "react-router": "^4.1.1",
+    "react-test-renderer": "16.0.0-alpha.6",
+    "redux": "^3.6.0",
+    "redux-actions": "^2.0.2",
+    "redux-logger": "^3.0.1",
+    "redux-thunk": "^2.2.0",
+    "supercluster": "^2.3.0",
+    "underscore": "^1.8.3"
+  },
+  "devDependencies": {
+    "babel-jest": "19.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react-native": "1.9.1",
+    "jest": "19.0.2"
+  },
+  "jest": {
+    "preset": "react-native"
+  }
 }
