@@ -1,20 +1,20 @@
-package com.sherpa_frontend;
+package com.sherpa;
 
 import android.app.Application;
-import android.util.Log;
 
 import com.facebook.react.ReactApplication;
+import com.cmcewen.blurview.BlurViewPackage;
+import com.wix.reactnativenotifications.RNNotificationsPackage;
+import fr.bamlab.rnimageresizer.ImageResizerPackage;
 import com.idehub.GoogleAnalyticsBridge.GoogleAnalyticsBridgePackage;
-import com.airbnb.android.react.maps.MapsPackage;
-import com.mapbox.reactnativemapboxgl.ReactNativeMapboxGLPackage;
-import com.babisoft.ReactNativeLocalization.ReactNativeLocalizationPackage;
-import com.i18n.reactnativei18n.ReactNativeI18n;
 import com.RNFetchBlob.RNFetchBlobPackage;
+import com.smixx.fabric.FabricPackage;
+import com.devialab.exif.RCTExifPackage;
 import com.learnium.RNDeviceInfo.RNDeviceInfo;
-import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
+import com.facebook.soloader.SoLoader;
 
 import java.util.Arrays;
 import java.util.List;
@@ -23,7 +23,7 @@ public class MainApplication extends Application implements ReactApplication {
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
     @Override
-    protected boolean getUseDeveloperSupport() {
+    public boolean getUseDeveloperSupport() {
       return BuildConfig.DEBUG;
     }
 
@@ -31,12 +31,13 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
+            new BlurViewPackage(),
+            new RNNotificationsPackage(),
+            new ImageResizerPackage(),
             new GoogleAnalyticsBridgePackage(),
-            new MapsPackage(),
-            new ReactNativeMapboxGLPackage(),
-            new ReactNativeLocalizationPackage(),
-            new ReactNativeI18n(),
             new RNFetchBlobPackage(),
+            new FabricPackage(),
+            new RCTExifPackage(),
             new RNDeviceInfo()
       );
     }
@@ -44,6 +45,12 @@ public class MainApplication extends Application implements ReactApplication {
 
   @Override
   public ReactNativeHost getReactNativeHost() {
-      return mReactNativeHost;
+    return mReactNativeHost;
+  }
+
+  @Override
+  public void onCreate() {
+    super.onCreate();
+    SoLoader.init(this, /* native exopackage */ false);
   }
 }
