@@ -1,6 +1,6 @@
 import React from 'react'
 import { TouchableOpacity, Text, View, Animated, Image,TouchableWithoutFeedback } from 'react-native'
-import notificationStyles from './styles/interactiveNotificationStyle'
+import notificationStyles from './styles/simpleErrorStyle'
 import { Fonts, Colors } from '../../../Themes/'
 
 export default class InteractiveNotification extends React.Component {
@@ -19,7 +19,9 @@ export default class InteractiveNotification extends React.Component {
     constructor(props){
         super(props);
         this.state={
-            offsetTop:new Animated.Value(-90)
+            offsetTop:new Animated.Value(-90),
+            opacity:new Animated.Value(1),
+            display:true
         }
     }
 
@@ -27,21 +29,29 @@ export default class InteractiveNotification extends React.Component {
         Animated.spring(this.state.offsetTop, {toValue: -90,friction:8}).start();
     }
 
+    hideFade(){
+        Animated.spring(this.state.opacity, {toValue: 0,friction:8}).start(()=>{
+            this.setState({display:false})
+        });
+    }
+
     show(){
+        this.setState({display:true})
+        Animated.timing(this.state.opacity, {toValue: 1,duration:0}).start();
         Animated.spring(this.state.offsetTop, {toValue: 0,friction:8}).start();
+        setTimeout(()=>{this.hideFade()},5000);
     }
 
 
     render () {
+        if(!this.state.display)return null;
         return (
-            <Animated.View style={[notificationStyles.errorContainer,{marginTop:this.state.offsetTop}]}>
                 <TouchableWithoutFeedback onPress={this.hide.bind(this)}>
-                    <View style={[notificationStyles.errorContainer,{backgroundColor:this.props.success?Colors.highlight:Colors.error}]}>
-                        <Image resizeMode="contain" style={notificationStyles.errorX} source={require('./../../../Images/icon-close.png')}></Image>
-                        <Text style={notificationStyles.errorMessage}>{this.props.notificationMessage.toUpperCase()}</Text>
-                    </View>
+                    <Animated.View style={[notificationStyles.errorContainer,{marginTop:this.state.offsetTop,backgroundColor:this.props.success?Colors.highlight:Colors.error,opacity:this.state.opacity}]}>
+                                <Image resizeMode="contain" style={notificationStyles.errorX} source={require('./../../../Images/icons/close-tooltipp.png')}></Image>
+                                <Text style={notificationStyles.errorMessage}>{this.props.notificationMessage.toUpperCase()}</Text>
+                    </Animated.View>
                 </TouchableWithoutFeedback>
-            </Animated.View>
         )
     }
 }
