@@ -40,6 +40,9 @@ class SherpaCameraRollPicker extends Component {
         if(typeof containerWidth != "undefined") {
             width = containerWidth;
         }
+
+        width=width-5;
+
         //console.log('component will mount');
         this._imageSize = (width - (imagesPerRow) ) / imagesPerRow;
 
@@ -130,7 +133,7 @@ class SherpaCameraRollPicker extends Component {
 
         var marker = selectedMarker ? selectedMarker :
             <Image
-                style={[styles.marker, {width: 25, height: 25, right: imageMargin + 5,bottom:imageMargin+5}]}
+                style={[styles.marker, {width: 30, height: 30, right: imageMargin -3 ,bottom:imageMargin-3}]}
                 source={require('../../../Images/icon-check-green.png')}
             />;
 
@@ -147,8 +150,9 @@ class SherpaCameraRollPicker extends Component {
                     <Image
                         source={{uri: item.node.image.uri}}
                         style={{height: this._imageSize, width: this._imageSize,opacity:isSelected?.7:1}} >
-                        { isSelected ? marker : null }
                     </Image>
+                    <View style={{position:'absolute',height: this._imageSize, width: this._imageSize,opacity:isSelected?1:0,backgroundColor:'rgba(255,255,255,.4)'}}></View>
+                    { isSelected ? marker : null }
                 </TouchableOpacity>
             </View>
         );
