@@ -3,6 +3,6 @@ import { Fonts, Colors } from '../../../../Themes/'
 
 export default StyleSheet.create({
     infoBoxContainer:{paddingTop:15,paddingBottom:0,paddingLeft:25,paddingRight:20,backgroundColor:'white'},
-    infoBoxTitle:{paddingBottom:10,fontSize:10,color:"#999999"},
-    infoBoxCopy:{fontSize:13}
+    infoBoxTitle:{paddingBottom:10,fontSize:10,letterSpacing:.6,color:"#999999",fontFamily:"TSTAR"},
+    infoBoxCopy:{fontSize:12,fontFamily:"Akkurat",fontWeight:"400",lineHeight:17,letterSpacing:.4}
 })
