@@ -9,6 +9,7 @@ import {
     ScrollView,
     TextInput,
     Alert
+
 } from 'react-native';
 import React, { Component } from 'react';
 import SimpleButton from '../../components/simpleButton';
@@ -32,7 +33,7 @@ class EditTripName extends React.Component {
             let shouldAdd=true;
             for(var j=0;j<props.deselectedMomentIDs.length;j++){
 
-                console.log(props.deselectedMomentIDs[j])
+                //console.log(props.deselectedMomentIDs[j])
                 if(props.momentData[i].id==props.deselectedMomentIDs[j]){
                     shouldAdd=false;
                 }
@@ -46,7 +47,7 @@ class EditTripName extends React.Component {
 
 
         this.state={
-            remainingCharacters:30,
+            remainingCharacters:20,
             text:props.tripData?props.tripData.name:"",
             positionBottom:new Animated.Value(14),
             tripData:null,
@@ -55,13 +56,16 @@ class EditTripName extends React.Component {
             momentData:props.momentData,
             deselectedMomentIDs:props.deselectedMomentIDs,
             canSubmit:true,
-            displayData
+            displayData,
+            maxLength:30
         }
 
     }
 
     componentDidMount(){
-        console.log(this.state.displayData);
+        //console.log(this.state.displayData);
+        console.log(this.state.deselectedMomentIDs)
+        console.log(this.state.displayData)
     }
 
     navActionRight(){
@@ -71,17 +75,20 @@ class EditTripName extends React.Component {
             this.refs.tripnameError.show();
         }else if(this.state.canSubmit){
             this.refs.tripnameError.hide();
-            console.log('submit submit')
+            //console.log('submit submit')
 
             this.setState({canSubmit:false})
             var moments=[];
             var momentIDs=[];
+            let addMomentsRaw=[];
+            let addMomentsFinal=[];
 
             //create moments or edit moments
             for(var i=0;i<this.state.displayData.length;i++){
-                //console.log('moment[i]',this.state.displayData[i]);
-                let momentPromise = createMoment(this.state.displayData[i]);
-                moments.push(momentPromise)
+                if(this.state.displayData[i].id.toString().indexOf("sherpa-internal")>-1){
+                    addMomentsRaw.push(this.state.displayData[i]);
+                    moments.push(createMoment(this.state.displayData[i]))
+                }
             }
 
 
@@ -89,25 +96,29 @@ class EditTripName extends React.Component {
             this.props.headerProgress.startToMiddle();
 
 
+
+
             getTripLocation(this.state.displayData).then((tripLocation)=>{
 
                 Promise.all(moments).then((momentsRes)=>{
                     var momentUploads=[];
                     var dates=[];
+
                     var coverMomentID="";
-                    for(var i=0;i<  this.state.displayData.length;i++){
+                    console.log(addMomentsRaw)
+                    for(var i=0;i<  addMomentsRaw.length;i++){
                             momentIDs.push(momentsRes[i].id)
                             if(this.state.displayData[i].isCover){
-                                console.log('cover is ',momentsRes[i])
+                                //console.log('cover is ',momentsRes[i])
                                 coverMomentID=momentsRes[i].id;
                             }
                             dates.push(this.state.displayData[i].date || new Date().getTime()/1000)
-                            if(this.state.displayData[i].id.toString().indexOf("sherpa-internal")>-1&&this.state.displayData[i].service=='sherpa-ios'){
+                            if(addMomentsRaw[i].id.toString().indexOf("sherpa-internal")>-1&&addMomentsRaw[i].service=='sherpa-ios'){
                                 momentUploads.push(uploadMoment(this.state.displayData[i],momentsRes[i]));
-                                console.log('add to moment uploads')
                             }
                     }
 
+                    console.log('final new moment ids',momentIDs);
                     dates.sort(function(a,b) {
                         return a - b
                     });
@@ -126,15 +137,21 @@ class EditTripName extends React.Component {
                         //    coverMomentID,
                         //    trip: this.props.tripData
                         //},'trip location::',tripLocation);
-                        console.log("cover moment id",coverMomentID);
-                        console.log("create trip!!!!!")
+
+                        console.log('uploaded moments',res)
+
+                        //console.log("cover moment id",coverMomentID);
+                        //console.log("create trip!!!!!")
+
                         createTrip({
                             momentIDs,
                             name: this.state.text,
                             startDate,
                             endDate,
                             coverMomentID,
-                            trip: this.props.tripData
+                            trip: this.props.tripData,
+                            addMomentIDs:momentIDs,
+                            removeMomentIDs:this.state.deselectedMomentIDs
                         }, tripLocation).then(()=> {
                             this.props.headerProgress.showSuccess();
                             setTimeout(this.props.refreshCurrentScene,500)
@@ -143,7 +160,7 @@ class EditTripName extends React.Component {
                             this.setState({canSubmit:true})
                             this.props.headerProgress.showError();
                         })
-                    }).catch((err)=>{console.log('err from upload')});
+                    }).catch((err)=>{console.log('err from upload',err)});
                 });
 
                 //console.log(this.props.navigator.getCurrentRoutes())
@@ -165,8 +182,10 @@ class EditTripName extends React.Component {
     render(){
         return(
             <View style={{backgroundColor:'white',width:SCREEN_WIDTH,height:SCREEN_HEIGHT,justifyContent:"center",alignItems:"center",flexDirection:"column"}}>
+
+
                 <TextInput clearTextOnFocus={true}
-                           clearButtonMode="while-editing"
+                           clearButtonMode="never"
                            style={
                                 {
                                     fontSize:35,
@@ -177,18 +196,31 @@ class EditTripName extends React.Component {
                                     marginTop:-190,
                                     textAlign:"center",
                                     letterSpacing:1,
-                                    width:SCREEN_WIDTH,
+                                    marginLeft:15,
+                                    width:SCREEN_WIDTH-30,
                                     marginBottom:10,
-                                    height:50,
-                                    alignItems:"center"
+                                    //height:50,
+                                    alignItems:"center",
+                                    height: Math.max(35, this.state.height)
                                 }
                             }
-                                      maxLength={30}
+                                      maxLength={this.state.maxLength}
+                           autoCapitalize='characters'
+                           multiline={true}
+                           onChange={(event) => {
+                              this.setState({
+                                    text: event.nativeEvent.text,
+                                    height: event.nativeEvent.contentSize.height,
+                                  });
+                           }}
                            onFocus={this.onStart.bind(this)}
                            onBlur={this.onEnd.bind(this)}
-                           onChangeText={(text) =>{this.setState({text:text.toUpperCase(),remainingCharacters:30-text.length})}}
-                                      value={this.state.text}
-                                      placeholder={'YOUR ALBUM'}
+
+                           placeholderTextColor="#cdcdcd"
+                           onChangeText={(text) =>{this.setState({text:text.toUpperCase(),remainingCharacters:this.state.maxLength-text.length})}}
+                                      value={this.state.text.toUpperCase()}
+                           placeholder={'YOUR ALBUM'}
+
                 >
                 </TextInput>
                 <Text style={{color:"#999999",fontFamily:Fonts.type.headline,fontSize:10,letterSpacing:.5}}>EDIT NAME HERE ({this.state.remainingCharacters})</Text>
